// projects/ui/src/lib/services/po-dialog/interfaces/po-dialog-confirm-literals.interface.ts
/**
 *  @usedBy PoDialogService
 *
 *  @description
 *
 * Interface para definição das literais usadas no serviço `po-dialog` para o tipo confirmação.
 */
export interface PoDialogConfirmLiterals {
  /** Rótulo do botão de "Cancelar". */
  cancel?: string;

  /** Rótulo do botão de "Confirmar". */
  confirm?: string;
}